+++ backend/vnc/settings.go
@@ -1,10 +1,16 @@
 package vnc
 
 type SettingsData struct {
-	ViewOnly         bool `json:"viewOnly"`
-	QualityLevel     int  `json:"qualityLevel"`
-	CompressionLevel int  `json:"compressionLevel"`
-	ShowDotCursor    bool `json:"showDotCursor"`
+	ViewOnly         bool    `json:"viewOnly"`
+	QualityLevel     int     `json:"qualityLevel"`
+	CompressionLevel int     `json:"compressionLevel"`
+	ShowDotCursor    bool    `json:"showDotCursor"`
+	Scaling          Scaling `json:"scaling"`
+}
+
+type Scaling struct {
+	ClipToWindow bool   `json:"clipToWindow"`
+	Resize       string `json:"resize"`
 }
 
 func getDefaultSettings() SettingsData {
@@ -13,5 +19,9 @@ func getDefaultSettings() SettingsData {
 		CompressionLevel: 2,
 		ShowDotCursor:    false,
 		ViewOnly:         false,
+		Scaling: Scaling{
+			ClipToWindow: false,
+			Resize:       "scale",
+		},
 	}
 }